Extract content readers into `app.toolkit`
