+++ app/app/files/domain/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from .content_metadata import ContentMetadata, Exif
+from .content_metadata import ContentMetadata
 from .file import File, MountedFile
 from .file_member import FileMember
 from .file_pending_deletion import FilePendingDeletion
@@ -12,7 +12,6 @@ from .shared_link import SharedLink
 
 __all__ = [
     "ContentMetadata",
-    "Exif",
     "File",
     "FileMember",
     "FilePendingDeletion",
