+++ tests/toolkit/readers/test_init.py
@@ -6,7 +6,7 @@ from unittest import mock
 
 import pytest
 
-from app.app.blobs.services.metadata import readers
+from app.toolkit import metadata
 
 pytestmark = [pytest.mark.anyio]
 
@@ -14,19 +14,19 @@ pytestmark = [pytest.mark.anyio]
 class TestLoad:
     async def test_image(self, image_content_with_exif: IO[bytes]):
         # GIVEN
-        target_guess = "app.app.blobs.services.metadata.readers.mediatypes.guess"
-        target_load = "app.app.blobs.services.metadata.readers.load_image_data"
+        target_guess = "app.toolkit.metadata.mediatypes.guess"
+        target_load = "app.toolkit.metadata.load_image_data"
         # WHEN
         with (
             mock.patch(target_guess, return_value="image/jpeg") as guess_mock,
             mock.patch(target_load) as load_image_mock,
         ):
-            result = await readers.load(image_content_with_exif)
+            result = await metadata.load(image_content_with_exif)
         # THEN
         assert result == load_image_mock.return_value
         guess_mock.assert_called_once_with(image_content_with_exif)
         load_image_mock.assert_called_once_with(image_content_with_exif)
 
     async def test_when_mediatype_is_not_supported(self):
-        data = await readers.load(BytesIO(b"Hello, world"))
+        data = await metadata.load(BytesIO(b"Hello, world"))
         assert data is None
